feat: Create basic layout scaffold
