+++ app/root.tsx
@@ -6,7 +6,7 @@ import {
   ScrollRestoration,
 } from "@remix-run/react";
 
-import "./tailwind.css";
+import "~/styles/tailwind.css";
 
 export default function App() {
   return (
